+++ opentelemetry-api/setup.py
@@ -44,7 +44,7 @@ setuptools.setup(
     include_package_data=True,
     long_description=open("README.rst").read(),
     long_description_content_type="text/x-rst",
-    install_requires=["typing; python_version<'3.5'"],
+    install_requires=["typing; python_version<'3.5'", "aiocontextvars"],
     extras_require={},
     license="Apache-2.0",
     package_dir={"": "src"},
@@ -58,9 +58,12 @@ setuptools.setup(
     zip_safe=False,
     entry_points={
         "opentelemetry_context": [
-            "default_context = "
-            "opentelemetry.context.default_context:"
-            "DefaultRuntimeContext",
+            "contextvars_context = "
+            "opentelemetry.context.contextvars_context:"
+            "ContextVarsRuntimeContext",
+            "threadlocal_context = "
+            "opentelemetry.context.threadlocal_context:"
+            "ThreadLocalRuntimeContext",
         ]
     },
 )
