--- opentelemetry-api/src/opentelemetry/context/default_context.py
@@ -1,34 +1,0 @@
-# Copyright 2020, OpenTelemetry Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-from opentelemetry.context.context import Context, RuntimeContext
-
-
-class DefaultRuntimeContext(RuntimeContext):
-    """A default implementation of the RuntimeContext interface using
-    a dictionary to store values.
-    """
-
-    def __init__(self) -> None:
-        self._current_context = Context()
-
-    def set_current(self, context: Context) -> None:
-        """See `opentelemetry.context.RuntimeContext.set_current`."""
-        self._current_context = context
-
-    def get_current(self) -> Context:
-        """See `opentelemetry.context.RuntimeContext.get_current`."""
-        return self._current_context
-
-
-__all__ = ["DefaultRuntimeContext"]
