+++ opentelemetry-api/tests/context/test_threadlocal_context.py
@@ -16,9 +16,7 @@ import unittest
 from unittest.mock import patch
 
 from opentelemetry import context
-from opentelemetry.sdk.context.threadlocal_context import (
-    ThreadLocalRuntimeContext,
-)
+from opentelemetry.context.threadlocal_context import ThreadLocalRuntimeContext
 
 
 def do_work() -> None:
@@ -33,7 +31,7 @@ class TestThreadLocalContext(unittest.TestCase):
         context.set_current(self.previous_context)
 
     @patch(
-        "opentelemetry.context._RUNTIME_CONTEXT", ThreadLocalRuntimeContext()
+        "opentelemetry.context._RUNTIME_CONTEXT", ThreadLocalRuntimeContext()  # type: ignore
     )
     def test_context(self):
         self.assertIsNone(context.get_value("say"))
@@ -51,7 +49,7 @@ class TestThreadLocalContext(unittest.TestCase):
         self.assertEqual(context.get_value("say", context=third), "bar")
 
     @patch(
-        "opentelemetry.context._RUNTIME_CONTEXT", ThreadLocalRuntimeContext()
+        "opentelemetry.context._RUNTIME_CONTEXT", ThreadLocalRuntimeContext()  # type: ignore
     )
     def test_set_value(self):
         first = context.set_value("a", "yyy")
